refactor(engine): split autoscaler.go into topic files
autoscaler.go mixed the reconcile loop with agent lifecycle, drift
cleanup and the billing-dependent teardown policy, making each concern
hard to find and change in isolation.

Move each concern into its own file: agents.go (lifecycle), cleanup.go
(provider/woodpecker drift), billing.go (teardown policy). autoscaler.go
now holds only the reconcile orchestration.

Behavior is unchanged. The duplicated idle-timeout check shared by the
drain and idle paths is folded into a single idleLongEnough helper, and
taskMatchesAgent documents the scope seam where server-injected scope
labels and internal-label stripping (and future user/org pool scopes)
will plug in.
