+++ engine/autoscaler.go
@@ -3,18 +3,25 @@ package engine
 import (
 	"context"
 	"fmt"
-	"strings"
-	"time"
 
 	"github.com/rs/zerolog/log"
 
 	"go.woodpecker-ci.org/autoscaler/config"
 	"go.woodpecker-ci.org/autoscaler/engine/types"
 	"go.woodpecker-ci.org/autoscaler/server"
-	"go.woodpecker-ci.org/autoscaler/utils"
 	"go.woodpecker-ci.org/woodpecker/v3/woodpecker-go/woodpecker"
 )
 
+// Autoscaler reconciles a pool of woodpecker agents against the current queue
+// state. The moving parts are split across the engine package by topic:
+//
+//   - autoscaler.go (this file): the reconcile loop that ties everything
+//     together.
+//   - buckets.go / labels.go: how tasks and agents are grouped and matched.
+//   - plan.go: how much each bucket should scale up or down.
+//   - agents.go: bringing agents up, draining, and removing them.
+//   - billing.go: the billing-model-dependent teardown policy.
+//   - cleanup.go: reconciling drift between the provider and woodpecker.
 type Autoscaler struct {
 	client               server.Client
 	agents               map[string]*woodpecker.Agent
@@ -39,286 +46,6 @@ func NewAutoscaler(ctx context.Context, p types.Provider, client server.Client, 
 		agents:               make(map[string]*woodpecker.Agent),
 		providerCapabilities: caps,
 	}, nil
-}
-
-// inTeardownWindow reports whether the agent is currently within the teardown
-// window before one of its paid-hour boundaries (anchored at its creation
-// time). Agents that have not reported a creation time are never in the window.
-func (a *Autoscaler) inTeardownWindow(agent *woodpecker.Agent) bool {
-	if agent.Created == 0 {
-		return false
-	}
-
-	age := time.Since(time.Unix(agent.Created, 0))
-	if age < 0 {
-		return false
-	}
-
-	window := a.config.AgentBillingTeardownMargin + a.config.ReconciliationInterval
-	// A window covering a whole hour (or more) means every moment qualifies.
-	if window >= time.Hour {
-		return true
-	}
-
-	return age%time.Hour >= time.Hour-window
-}
-
-func (a *Autoscaler) loadAgents(_ context.Context) error {
-	a.agents = make(map[string]*woodpecker.Agent)
-
-	agents, err := a.client.AgentList()
-	if err != nil {
-		return fmt.Errorf("client.AgentList: %w", err)
-	}
-
-	prefix := fmt.Sprintf("pool-%s-agent-", a.config.PoolID)
-	for _, agent := range agents {
-		if strings.HasPrefix(agent.Name, prefix) {
-			a.agents[agent.Name] = agent
-		}
-	}
-
-	return nil
-}
-
-// createAgents deploys `amount` new agents into the given bucket.
-// It first tries to reactivate matching no-schedule agents
-// (which are already provisioned on the provider) before deploying new ones.
-func (a *Autoscaler) createAgents(ctx context.Context, bucket agentBucket, amount int) error {
-	const suffixLength = 4
-
-	if amount <= 0 {
-		return nil
-	}
-
-	// First, reactivate matching no-schedule agents.
-	reactivated := 0
-	for _, agent := range a.agents {
-		if reactivated >= amount {
-			break
-		}
-		if !agent.NoSchedule {
-			continue
-		}
-		if !agentMatchesCapability(agent, bucket.Capability) {
-			continue
-		}
-
-		log.Info().Str("agent", agent.Name).Msg("reactivate agent")
-		agent.NoSchedule = false
-		if _, err := a.client.AgentUpdate(agent); err != nil {
-			return fmt.Errorf("client.AgentUpdate: %w", err)
-		}
-		reactivated++
-	}
-
-	// Deploy fresh agents for whatever's left.
-	for i := reactivated; i < amount; i++ {
-		agent, err := a.client.AgentCreate(&woodpecker.Agent{
-			Name: fmt.Sprintf("pool-%s-agent-%s", a.config.PoolID, utils.RandomString(suffixLength)),
-		})
-		if err != nil {
-			return fmt.Errorf("client.AgentCreate: %w", err)
-		}
-
-		log.Info().
-			Str("agent", agent.Name).
-			Str("platform", bucket.Capability.Platform).
-			Str("backend", string(bucket.Capability.Backend)).
-			Msg("deploying agent")
-
-		if err := a.provider.DeployAgent(ctx, agent, bucket.Capability); err != nil {
-			return fmt.Errorf("types.DeployAgent: %w", err)
-		}
-
-		a.agents[agent.Name] = agent
-	}
-
-	return nil
-}
-
-func (a *Autoscaler) drainAgents(_ context.Context, bucket agentBucket, amount int) error {
-	if amount <= 0 {
-		return nil
-	}
-	drained := 0
-	for _, agent := range a.agents {
-		if drained >= amount {
-			break
-		}
-		// agent is already marked for draining
-		if agent.NoSchedule {
-			continue
-		}
-		// only drain agents that belong to this bucket
-		if !agentMatchesCapability(agent, bucket.Capability) {
-			continue
-		}
-		// agent has never contacted the server => not ready for draining
-		if agent.LastContact == 0 {
-			continue
-		}
-
-		if a.config.BillingModel == types.BillingHourlyRoundUp {
-			// hourly-round-up: the hour is already paid for, so keep the
-			// agent schedulable until just before its hour boundary even
-			// while idle, then drain it inside the teardown window.
-			if !a.inTeardownWindow(agent) {
-				continue
-			}
-		} else if time.Since(time.Unix(agent.LastWork, 0)) < a.config.AgentIdleTimeout {
-			// agent has recently done work => not ready for draining
-			continue
-		}
-
-		log.Info().Str("agent", agent.Name).Msg("drain agent")
-		agent.NoSchedule = true
-		if _, err := a.client.AgentUpdate(agent); err != nil {
-			return fmt.Errorf("client.AgentUpdate: %w", err)
-		}
-		drained++
-	}
-
-	return nil
-}
-
-func (a *Autoscaler) isAgentIdle(agent *woodpecker.Agent) (bool, error) {
-	tasks, err := a.client.AgentTasksList(agent.ID)
-	if err != nil {
-		return false, fmt.Errorf("client.AgentTasksList: %w", err)
-	}
-
-	// agent still has tasks => not idle
-	if len(tasks) > 0 {
-		return false, nil
-	}
-
-	// hourly-round-up: recency of work does not gate removal. The paid hour is
-	// kept warm by the drain stage; once an agent is eligible for removal the
-	// only thing that protects it is an in-flight task (checked above).
-	if a.config.BillingModel == types.BillingHourlyRoundUp {
-		return true, nil
-	}
-
-	// agent has done work recently => not idle
-	if time.Since(time.Unix(agent.LastWork, 0)) < a.config.AgentIdleTimeout {
-		return false, nil
-	}
-
-	return true, nil
-}
-
-func (a *Autoscaler) removeAgent(ctx context.Context, agent *woodpecker.Agent, reason string) error {
-	isIdle, err := a.isAgentIdle(agent)
-	if err != nil {
-		return err
-	}
-	if !isIdle {
-		log.Info().Str("agent", agent.Name).Msg("agent is still processing workload")
-		return nil
-	}
-
-	log.Info().Str("agent", agent.Name).Str("reason", reason).Msgf("removing agent")
-
-	err = a.provider.RemoveAgent(ctx, agent)
-	if err != nil {
-		return err
-	}
-
-	err = a.client.AgentDelete(agent.ID)
-	if err != nil {
-		return fmt.Errorf("client.AgentDelete: %w", err)
-	}
-
-	delete(a.agents, agent.Name)
-
-	return nil
-}
-
-func (a *Autoscaler) removeDrainedAgents(ctx context.Context) error {
-	for _, agent := range a.agents {
-		if !agent.NoSchedule {
-			continue
-		}
-
-		// hourly-round-up: a drained agent that rolled into a fresh paid hour
-		// (e.g. it was busy at the boundary) stays up until its next teardown
-		// window rather than wasting the hour just bought.
-		if a.config.BillingModel == types.BillingHourlyRoundUp && !a.inTeardownWindow(agent) {
-			continue
-		}
-
-		err := a.removeAgent(ctx, agent, "was drained")
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (a *Autoscaler) cleanupDanglingAgents(ctx context.Context) error {
-	providerAgentNames, err := a.provider.ListDeployedAgentNames(ctx)
-	if err != nil {
-		return err
-	}
-
-	// Build the provider-side set up front so the two reconciliation
-	// directions below are independent of each other.
-	providerSet := make(map[string]struct{}, len(providerAgentNames))
-	for _, name := range providerAgentNames {
-		providerSet[name] = struct{}{}
-	}
-
-	// On provider but not on woodpecker → tear down on the provider.
-	for name := range providerSet {
-		if _, ok := a.agents[name]; ok {
-			continue
-		}
-		log.Info().Str("agent", name).Str("reason", "not found on woodpecker").Msg("remove agent")
-		if err := a.provider.RemoveAgent(ctx, &woodpecker.Agent{Name: name}); err != nil {
-			return fmt.Errorf("types.RemoveAgent: %w", err)
-		}
-	}
-
-	// On woodpecker but not on provider → delete on woodpecker.
-	for name, agent := range a.agents {
-		if _, ok := providerSet[name]; ok {
-			continue
-		}
-		log.Info().Str("agent", name).Str("reason", "not found on provider").Msg("remove agent")
-		if err := a.client.AgentDelete(agent.ID); err != nil {
-			return fmt.Errorf("client.AgentDelete: %w", err)
-		}
-		delete(a.agents, name)
-	}
-
-	return nil
-}
-
-func (a *Autoscaler) cleanupStaleAgents(ctx context.Context) error {
-	// remove agents that haven't contacted the server for a while (including agents that never contacted the server)
-	for _, agent := range a.agents {
-		if agent.NoSchedule {
-			continue
-		}
-
-		lastContact := agent.LastContact
-
-		// if agent has never contacted the server, use the creation time
-		if lastContact == 0 {
-			lastContact = agent.Created
-		}
-
-		if time.Since(time.Unix(lastContact, 0)) > a.config.AgentInactivityTimeout {
-			err := a.removeAgent(ctx, agent, "hasn't connected to the server for a while")
-			if err != nil {
-				return err
-			}
-		}
-	}
-
-	return nil
 }
 
 // Reconcile periodically checks the status of the agent pool and adjusts
